requirements.txt
biopython==1.83
CairoSVG==2.7.1
keras==2.10.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.3.2
stellargraph==1.2.1
tensorflow==2.10.0
tqdm==4.67.1
pyBigWig==0.3.22
pyfaidx==0.8.1.3
scipy==1.10.1
matplotlib==3.7.5
gseapy==1.1.6
